Add files via upload
The task was to complete a project as a group in my project management class. We were asked to split the project requirements amongst ourselves and it was our own responsibility to get together outside of class to complete the project. I took lead and assigned everyone roles to complete and gather information. I created the entire power point and am responsible for the information/data in pages 8-12. I received a 100 for my contribution to the project and for taking on extra slides since one of our members did not respond at all.